pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trillium-watts"
version = "1.0.0"
description = "Energy demand forecasting and solar simulation for Leticia, Amazonas, Colombia"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "tensorflow>=2.14",
    "statsmodels>=0.14",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "plotly>=5.15",
    "streamlit>=1.28",
    "gdown>=5.0",
    "pyyaml>=6.0",
    "joblib>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
]
app = [
    "pandas>=2.0",
    "numpy>=1.24",
    "plotly>=5.15",
    "streamlit>=1.28",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
where = ["src"]
